Allow multiple calls to SetCacheControl
Suppose server method A calls method B and then sets a CacheControl.
If method B also sets a CacheControl, then it would be written to the
gRPC trailer and method A's SetCacheControl would fail to take effect
(since the trailer was already written). This makes it hard to compose
server methods.

This commit defers setting the CacheControl trailer until the request
handling is complete. It means, unfortunately, that server impls must
also be wrapped in a codegenned wrapper (CachedXyzServer). This
wrapper must be the outermost server type.